Rename "guesture" to "gesture", fix spelling mistake
