+++ app/shared/src/commonMain/kotlin/ui/subject/episode/EpisodeVideo.kt
@@ -60,14 +60,14 @@ import me.him188.ani.app.videoplayer.ui.PlayerControllerState
 import me.him188.ani.app.videoplayer.ui.VideoPlayer
 import me.him188.ani.app.videoplayer.ui.VideoScaffold
 import me.him188.ani.app.videoplayer.ui.VideoSideSheetsController
-import me.him188.ani.app.videoplayer.ui.guesture.GestureFamily
-import me.him188.ani.app.videoplayer.ui.guesture.GestureLock
-import me.him188.ani.app.videoplayer.ui.guesture.LevelController
-import me.him188.ani.app.videoplayer.ui.guesture.LockableVideoGestureHost
-import me.him188.ani.app.videoplayer.ui.guesture.ScreenshotButton
-import me.him188.ani.app.videoplayer.ui.guesture.mouseFamily
-import me.him188.ani.app.videoplayer.ui.guesture.rememberGestureIndicatorState
-import me.him188.ani.app.videoplayer.ui.guesture.rememberSwipeSeekerState
+import me.him188.ani.app.videoplayer.ui.gesture.GestureFamily
+import me.him188.ani.app.videoplayer.ui.gesture.GestureLock
+import me.him188.ani.app.videoplayer.ui.gesture.LevelController
+import me.him188.ani.app.videoplayer.ui.gesture.LockableVideoGestureHost
+import me.him188.ani.app.videoplayer.ui.gesture.ScreenshotButton
+import me.him188.ani.app.videoplayer.ui.gesture.mouseFamily
+import me.him188.ani.app.videoplayer.ui.gesture.rememberGestureIndicatorState
+import me.him188.ani.app.videoplayer.ui.gesture.rememberSwipeSeekerState
 import me.him188.ani.app.videoplayer.ui.hasPageAsState
 import me.him188.ani.app.videoplayer.ui.progress.AudioSwitcher
 import me.him188.ani.app.videoplayer.ui.progress.MediaProgressIndicatorText
