+++ app/shared/video-player/src/androidMain/kotlin/ui/PlayerGestureHost.android.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 OpenAni and contributors.
+ * Copyright (C) 2024-2025 OpenAni and contributors.
  *
  * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
@@ -17,8 +17,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import me.him188.ani.app.videoplayer.ui.guesture.GestureIndicator
-import me.him188.ani.app.videoplayer.ui.guesture.rememberGestureIndicatorState
+import me.him188.ani.app.videoplayer.ui.gesture.GestureIndicator
+import me.him188.ani.app.videoplayer.ui.gesture.rememberGestureIndicatorState
 
 
 @Composable
